requirements.txt
torch===2.2.0
torchaudio===2.2.0
torchvision===0.1.6
transformer-lens===1.14.0
tqdm===4.66.1
numpy===1.26.3
python-dotenv===1.0.1
wandb===0.16.2
einops===0.7.0
plotly===5.18.0
matplotlib===3.8.2
pytest===8.0.0